Allow deleting multiple states at once
On the backend, also add possibility for ItemActions to be applied to
multiple indices. If they are marked in this way, they will be given a
list of indices instead of a single one.

Also adds the Del shortcut to delete states.

This is apparently a pretty popular feature request.